Move DataRef to expr package
