+++ FortranAst/src/pt/up/fe/specs/fortran/ast/nodes/expr/DataRef.java
@@ -1,10 +1,9 @@
-package pt.up.fe.specs.fortran.ast.nodes.variable;
+package pt.up.fe.specs.fortran.ast.nodes.expr;
 
 import org.suikasoft.jOptions.Datakey.DataKey;
 import org.suikasoft.jOptions.Datakey.KeyFactory;
 import org.suikasoft.jOptions.Interfaces.DataStore;
 import pt.up.fe.specs.fortran.ast.nodes.FortranNode;
-import pt.up.fe.specs.fortran.ast.nodes.expr.Designator;
 
 import java.util.Collection;
